extend the editor prefix to the whole file
this limits all the selectors to be applied just within the editor
prefix. this makes this file safe to coexist with other css files. i
update the indentation in order to reflect the effect of this
action. many of the included rules here come directly from the mock. i
plan to remove the redundancies with time.